% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrodesign.R
\name{retrodesign}
\alias{retrodesign}
\title{Post hoc power computation}
\usage{
retrodesign(A, SE = 1, sig.level = 0.05, df = Inf, B = 10000)
}
\arguments{
\item{A}{numeric vector with effect size.}

\item{SE}{numeric with standard error (default=1).}

\item{sig.level}{significance level (default=0.05).}

\item{df}{degrees of freedom (default=infinity).}

\item{B}{number of simulation replications used for typeM computation (default=10000).}
}
\value{
A data frame with variables
\describe{
\item{\code{power}}{power of test}
\item{\code{typeS}}{risk of Type S error}
\item{\code{exaggeration}}{average exaggeration ratio}
}
}
\description{
\code{retrodesign} performs post hoc power analysis as proposed by Gelman & Carlin.
}
\note{
R code adapted from Gelman & Carlin (2014).
}
\examples{
mydata <- retrodesign(c(seq(0,1,0.01),seq(1,10,0.1),100),1)

# Type S error
library(ggplot2)
ggplot(mydata,aes(x=power,y=typeS)) + geom_line() + xlab("Power") + ylab("Type S error")

# Type M error
ggplot(mydata,aes(x=power,y=exaggeration)) + geom_line() + xlab("Power") + ylab("Exaggeration ratio") + ylim(c(0,12)) + geom_abline(slope=0,intercept=1,linetype=2)

}
\references{
Gelman & Carlin: "Beyond Power Calculations: Assessing Type S (Sign) and Type M (Magnitude) Errors", Perspectives on Psychological Science, 1-11, 2014.
}
\author{
Bo Markussen
}
\keyword{htest}
