% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_designDiagram.R
\name{designDiagram-class}
\alias{designDiagram-class}
\alias{print.designDiagram}
\alias{summary.designDiagram}
\alias{plot.designDiagram}
\title{The \code{designDiagram} class and some basic methods}
\usage{
\method{print}{designDiagram}(x, ...)

\method{summary}{designDiagram}(x, ...)

\method{plot}{designDiagram}(
  x,
  circle = "none",
  pvalue = (circle == "MSS"),
  kill.intercept = TRUE,
  color = ifelse(circle == "MSS", "lightblue", "lightgreen"),
  boldface = FALSE,
  arrow.type = arrow(angle = 20),
  r.scaling = 1,
  xlim = c(0, 1),
  ylim = c(0, 1),
  fontsize.horizontal = 0.03,
  fontsize.vertical = 0.05,
  horizontal = FALSE
)
}
\arguments{
\item{x}{object of class \code{designDiagram}}

\item{circle}{character specifying which circles to draw at the terms: \code{"none"}=no circles, \code{"SS"}=a circle with area proportional to the associated Sum-of-Squares, \code{"MSS"}=a circle with area proportional to the associated Mean-Sum-of-Squares. The two latter options are only available if a response variable was specified for the design. Defaults to \code{"none"}.}

\item{pvalue}{boolean specifying whether p-values should be inserted on the graphs. This is only possible if a response variable was specified. Defaults to \code{TRUE} is \code{circle="MSS"} and \code{FALSE} otherwise.}

\item{kill.intercept}{boolean specifying whether circle for the intercept should be removed. This is practicable since the intercept term otherwise may overweight the remaining terms. Defaults to \code{TRUE}.}

\item{color}{color of circles. Defaults to \code{"lightgreen"} for Sum-of-Squares and to \code{"lightblue"} for Mean-Sum-of-Squares.}

\item{arrow.type}{specifying arrow heads via \code{\link{arrow}}. Defaults to \code{arrow(angle=20,length=unit(4,"mm"))}.}

\item{xlim}{x-range of diagram plot. Defaults to \code{c(0,1)}.}

\item{ylim}{y-range of diagram plot. Defaults to \code{c(0,1)}.}

\item{horizontal}{boolen specifying if the design diagram should be drawn horizontally. Defauls to \code{FALSE}, which gives a vertical layout.}

\item{circle.scaling}{numeric specifying size scaling of circles. Defaults to \code{1}, which corresponds to the largest circle having a radius that is half of the shortest distance between two nodes.}
}
\description{
Objects of class \code{designDiagram} as generated by \code{\link{DD}} is a list with entries as specified below.
\describe{
\item{\code{terms}}{Named vector with all terms in the design.}
\item{\code{random.terms}}{Vector with the random terms in the design.}
\item{\code{Nparm}}{Named vector with the number of parameters for the terms.}
\item{\code{df}}{Named vector with the degrees of freedom for the terms.}
\item{\code{collinearities}}{Named vector with the number of (approximate) collinearities that have been removed from the terms.}
\item{\code{SS}}{Named vector with Sum-of-Squares if a response variable was specified.}
\item{\code{MSS}}{Named vector with Mean-Sum-of-Squares if a response variable was specified.}
\item{\code{relations}}{Named matrix with relations between variables with the following interpretation: "0"=linear indepent, "<"=row term is a subspace of column, "<-"=row term is a subspace of column term and no other terms are inbetween, ">" and "->" the similar interpretatioin between columns and rows, name=name of minimum between row and column term.}
\item{\code{pvalue}}{Named matrix with p-values for F-tests. p-values are stated at the collapsed nesting, but F-test are done against the most coarse nested random effect.}
\item{\code{inner}}{Named matrix of squared inner products of subspaces with nesting subspaces removed. Rouded at 6'th digits, and used to decide orthogonality of the design.}
\item{\code{response}}{Logical stating whether a response variable was present.}
}
}
\seealso{
\code{\link{DD}}
}
