% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_designDiagram.R
\name{designDiagram-class}
\alias{designDiagram-class}
\alias{print.designDiagram}
\alias{summary.designDiagram}
\alias{plot.designDiagram}
\title{The \code{designDiagram} class and some basic methods}
\usage{
\method{print}{designDiagram}(x, ...)

\method{summary}{designDiagram}(object, ...)

\method{plot}{designDiagram}(
  x,
  circle = "none",
  pvalue = (circle == "MSS"),
  kill = ~1,
  ca = NULL,
  max.area = NULL,
  relative = 0.01,
  color = ifelse(circle == "MSS", "lightblue", "lightgreen"),
  circle.scaling = 1,
  arrow.type = arrow(angle = 20, length = unit(4, "mm")),
  xlim = c(0, 1),
  ylim = c(0, 1),
  horizontal = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{designDiagram}}

\item{...}{not used.}

\item{object}{object of class \code{designDiagram}}

\item{circle}{character specifying which circles to draw at the terms: \code{"none"}=no circles, \code{"SS"}=a circle with area proportional to the associated Sum-of-Squares, \code{"MSS"}=a circle with area proportional to the associated Mean-Sum-of-Squares, \code{"III"}=TO BE DESCRIBED. The three latter options are only available if a response variable was specified for the design. Defaults to \code{"none"}.}

\item{pvalue}{boolean specifying whether p-values should be inserted on the graphs. This is only possible if a response variable was specified. Defaults to \code{TRUE} is \code{circle="MSS"} and \code{FALSE} otherwise.}

\item{kill}{formula specifying which cirlces not to plot. Defaults to \code{~1} corresponding to not plotting the intercept term (that otherwise may overweight the remaining terms).}

\item{ca}{boolean deciding whether collinearity analysis is visualized. If \code{NULL} then set \code{TRUE} for non-orthogonal designs, and to \code{FALSE} for orthogonal designs. Defaults to \code{NULL}.}

\item{max.area}{numeric specifying the used maximal area of circles. If \code{NULL} then \code{max.area} is derived from \code{SS} or \code{MSS} according to value of \code{circle}. Defaults to \code{NULL}.}

\item{relative}{positive numeric, which specifies needed relative increase for an area to be visualized in the collinearity analysis. Defaults to \code{0.01}.}

\item{color}{color of circles when \code{ca=FALSE}. Defaults to \code{"lightgreen"} for Sum-of-Squares and to \code{"lightblue"} for Mean-Sum-of-Squares.}

\item{circle.scaling}{numeric specifying size scaling of circles. Defaults to \code{1}, which corresponds to the largest circle having a radius that is half of the shortest distance between two nodes.}

\item{arrow.type}{specifying arrow heads via \code{\link[grid]{arrow}}. Defaults to \code{arrow(angle=20,length=unit(4,"mm"))}.}

\item{xlim}{x-range of diagram plot. Defaults to \code{c(0,1)}.}

\item{ylim}{y-range of diagram plot. Defaults to \code{c(0,1)}.}

\item{horizontal}{boolen specifying if the design diagram should be drawn horizontally or vertically. Defauls to \code{TRUE}.}
}
\description{
Objects of class \code{designDiagram} as generated by \code{\link{DD}} is a list with entries as specified below.
\describe{
\item{\code{terms}}{Named vector with all terms in the design.}
\item{\code{random.terms}}{Vector with the random terms in the design.}
\item{\code{Nparm}}{Named vector with the number of parameters for the terms.}
\item{\code{df}}{Named vector with the degrees of freedom for the terms.}
\item{\code{SS}}{Named matrix with Sum-of-Squares if a response variable was specified.}
\item{\code{MSS}}{Named matrix with Mean-Sum-of-Squares if a response variable was specified.}
\item{\code{relations}}{Named matrix with relations between variables with the following interpretation: "0"=linear indepent, "<"=row term is a subspace of column, "<-"=row term is a subspace of column term and no other terms are inbetween, ">" and "->" the similar interpretatioin between columns and rows, name=name of minimum between row and column term.}
\item{\code{pvalue}}{Named matrix with p-values for Type-I F-tests. p-values are stated at the collapsed nesting, but F-test are done against the most coarse nested random effect.}
\item{\code{inner}}{Named matrix of squared inner products of subspaces with nesting subspaces removed. Rouded at 6'th digits, and used to decide orthogonality of the design.}
\item{\code{response}}{Logical stating whether a response variable was present.}
}
}
\seealso{
\code{\link{DD}}
}
