% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans_ED.R
\name{emmeans_ED}
\alias{emmeans_ED}
\title{Make emmeans object for an expected dose}
\usage{
emmeans_ED(
  object,
  specs,
  left,
  right,
  tran = NULL,
  p = 0.5,
  p.name = "probability"
)
}
\arguments{
\item{object}{An object that can be given to \code{\link[emmeans]{emmeans}}. Typically a model fitted by \code{\link{glm}}.}

\item{specs}{As for \code{\link[emmeans]{emmeans}}. Typically as one-sided \code{\link{formula}}.}

\item{left}{A list specifying the left end point of the linear span of continuous variables in which to measure the ED values.}

\item{right}{A list specifying the right end point of the linear span of continuous variables in which to measure the ED values.}

\item{tran}{Possible transformation of the scale of the ED values. If given then backtransformation can be done using the technology of the \code{\link[emmeans]{emmeans}}. The default value \code{tran=NULL} corresponds to no transformation.}

\item{p}{Numeric vector given the targeted predictions. Typically probabilities, where the default value \code{p=0.5} corresponds to ED50.}

\item{p.name}{The name of the variable containing \code{p}. If \code{p} contains more than one value, then this will also appear in \code{@misc$by.vars} in the \code{emmGrid} object. Defaults to \code{p.name="probability"}.}
}
\description{
Solves linear equations in continuous explanatory variables in order to find the expected dose. A typical application could be to find LD50, i.e. the lethal dose killing 50 percent of the population, from a probit analysis fitted by \code{\link{glm}}. The associated variance-covariance matrix is found using the Delta method.
}
\author{
Bo Markussen
}
